moved fetch job method from actions to jobService(#2mky6cn)
